requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
insightface
opencv-python
pillow
pandas
numpy
pydantic>=2.0,<3
PyYAML>=6.0.1,<7
tqdm
safetensors
matplotlib
jinja2
pytest>=7.0,<9
